pure refactor off of master
