+++ mlflow/genai/judges/adapters/databricks_adapter.py
@@ -1,0 +1,439 @@
+"""Databricks adapter for judge model invocation with tool calling support."""
+
+from __future__ import annotations
+
+import json
+import logging
+import time
+from dataclasses import dataclass
+from typing import TYPE_CHECKING, Any
+
+import pydantic
+import requests
+
+if TYPE_CHECKING:
+    from mlflow.types.llm import ChatMessage
+
+from mlflow.entities.assessment import Feedback
+from mlflow.entities.assessment_source import AssessmentSource, AssessmentSourceType
+from mlflow.exceptions import MlflowException
+from mlflow.genai.judges.constants import _DATABRICKS_DEFAULT_JUDGE_MODEL
+from mlflow.genai.judges.utils.parsing_utils import _sanitize_justification
+from mlflow.genai.judges.utils.prompt_utils import _split_messages_for_databricks
+from mlflow.protos.databricks_pb2 import BAD_REQUEST, INVALID_PARAMETER_VALUE
+from mlflow.version import VERSION
+
+_logger = logging.getLogger(__name__)
+
+
+def _check_databricks_agents_installed() -> None:
+    """Check if databricks-agents is installed for databricks judge functionality.
+
+    Raises:
+        MlflowException: If databricks-agents is not installed.
+    """
+    try:
+        import databricks.agents.evals  # noqa: F401
+    except ImportError:
+        raise MlflowException(
+            f"To use '{_DATABRICKS_DEFAULT_JUDGE_MODEL}' as the judge model, the Databricks "
+            "agents library must be installed. Please install it with: "
+            "`pip install databricks-agents`",
+            error_code=BAD_REQUEST,
+        )
+
+
+def call_chat_completions(user_prompt: str, system_prompt: str) -> Any:
+    """
+    Invokes the Databricks chat completions API using the databricks.agents.evals library.
+
+    Args:
+        user_prompt (str): The user prompt.
+        system_prompt (str): The system prompt.
+
+    Returns:
+        The chat completions result.
+
+    Raises:
+        MlflowException: If databricks-agents is not installed.
+    """
+    _check_databricks_agents_installed()
+
+    from databricks.rag_eval import context, env_vars
+
+    env_vars.RAG_EVAL_EVAL_SESSION_CLIENT_NAME.set(f"mlflow-judge-optimizer-v{VERSION}")
+
+    @context.eval_context
+    def _call_chat_completions(user_prompt: str, system_prompt: str):
+        managed_rag_client = context.get_context().build_managed_rag_client()
+
+        return managed_rag_client.get_chat_completions_result(
+            user_prompt=user_prompt,
+            system_prompt=system_prompt,
+        )
+
+    return _call_chat_completions(user_prompt, system_prompt)
+
+
+def _parse_databricks_judge_response(
+    llm_output: str | None,
+    assessment_name: str,
+) -> Feedback:
+    """
+    Parse the response from Databricks judge into a Feedback object.
+
+    Args:
+        llm_output: Raw output from the LLM, or None if no response.
+        assessment_name: Name of the assessment.
+
+    Returns:
+        Feedback object with parsed results or error.
+    """
+    source = AssessmentSource(
+        source_type=AssessmentSourceType.LLM_JUDGE, source_id=_DATABRICKS_DEFAULT_JUDGE_MODEL
+    )
+    if not llm_output:
+        return Feedback(
+            name=assessment_name,
+            error="Empty response from Databricks judge",
+            source=source,
+        )
+    try:
+        response_data = json.loads(llm_output)
+    except json.JSONDecodeError as e:
+        return Feedback(
+            name=assessment_name,
+            error=f"Invalid JSON response from Databricks judge: {e}",
+            source=source,
+        )
+    if "result" not in response_data:
+        return Feedback(
+            name=assessment_name,
+            error=f"Response missing 'result' field: {response_data}",
+            source=source,
+        )
+    return Feedback(
+        name=assessment_name,
+        value=response_data["result"],
+        rationale=response_data.get("rationale", ""),
+        source=source,
+    )
+
+
+def _invoke_databricks_judge(
+    prompt: str | list["ChatMessage"],
+    assessment_name: str,
+) -> Feedback:
+    """
+    Invoke the Databricks judge using the databricks.agents.evals library.
+
+    Uses the direct chat completions API for clean prompt submission without
+    any additional formatting or template requirements.
+
+    Args:
+        prompt: The formatted prompt with template variables filled in.
+        assessment_name: The name of the assessment.
+
+    Returns:
+        Feedback object from the Databricks judge.
+
+    Raises:
+        MlflowException: If databricks-agents is not installed.
+    """
+    try:
+        if isinstance(prompt, str):
+            system_prompt = None
+            user_prompt = prompt
+        else:
+            prompts = _split_messages_for_databricks(prompt)
+            system_prompt = prompts.system_prompt
+            user_prompt = prompts.user_prompt
+
+        llm_result = call_chat_completions(user_prompt, system_prompt)
+        return _parse_databricks_judge_response(llm_result.output, assessment_name)
+
+    except Exception as e:
+        _logger.debug(f"Failed to invoke Databricks judge: {e}")
+        return Feedback(
+            name=assessment_name,
+            error=f"Failed to invoke Databricks judge: {e}",
+            source=AssessmentSource(
+                source_type=AssessmentSourceType.LLM_JUDGE,
+                source_id=_DATABRICKS_DEFAULT_JUDGE_MODEL,
+            ),
+        )
+
+
+@dataclass
+class InvokeDatabricksModelOutput:
+    response: str
+    request_id: str | None
+    num_prompt_tokens: int | None
+    num_completion_tokens: int | None
+
+
+def _parse_databricks_model_response(
+    res_json: dict[str, Any], headers: dict[str, Any]
+) -> InvokeDatabricksModelOutput:
+    """
+    Parse and validate the response from a Databricks model invocation.
+
+    Args:
+        res_json: The JSON response from the model
+        headers: The response headers
+
+    Returns:
+        InvokeDatabricksModelOutput with parsed response data
+
+    Raises:
+        MlflowException: If the response structure is invalid
+    """
+    # Validate and extract choices
+    choices = res_json.get("choices", [])
+    if not choices:
+        raise MlflowException(
+            "Invalid response from Databricks model: missing 'choices' field",
+            error_code=INVALID_PARAMETER_VALUE,
+        )
+
+    first_choice = choices[0]
+    if "message" not in first_choice:
+        raise MlflowException(
+            "Invalid response from Databricks model: missing 'message' field",
+            error_code=INVALID_PARAMETER_VALUE,
+        )
+
+    content = first_choice.get("message", {}).get("content")
+    if content is None:
+        raise MlflowException(
+            "Invalid response from Databricks model: missing 'content' field",
+            error_code=INVALID_PARAMETER_VALUE,
+        )
+
+    # Handle reasoning response (list of content items)
+    if isinstance(content, list):
+        text_content = None
+        for item in content:
+            if isinstance(item, dict) and item.get("type") == "text":
+                text_content = item.get("text")
+                break
+
+        if text_content is None:
+            raise MlflowException(
+                "Invalid reasoning response: no text content found in response list",
+                error_code=INVALID_PARAMETER_VALUE,
+            )
+        content = text_content
+
+    usage = res_json.get("usage", {})
+
+    return InvokeDatabricksModelOutput(
+        response=content,
+        request_id=headers.get("x-request-id"),
+        num_prompt_tokens=usage.get("prompt_tokens"),
+        num_completion_tokens=usage.get("completion_tokens"),
+    )
+
+
+def _invoke_databricks_model(
+    *,
+    model_name: str,
+    prompt: str,
+    num_retries: int,
+    response_format: type[pydantic.BaseModel] | None = None,
+) -> InvokeDatabricksModelOutput:
+    from mlflow.utils.databricks_utils import get_databricks_host_creds
+
+    # B-Step62: Why not use mlflow deployment client?
+    host_creds = get_databricks_host_creds()
+    api_url = f"{host_creds.host}/serving-endpoints/{model_name}/invocations"
+
+    # Implement retry logic with exponential backoff
+    last_exception = None
+    for attempt in range(num_retries + 1):
+        try:
+            # Build request payload
+            payload = {
+                "messages": [
+                    {
+                        "role": "user",
+                        "content": prompt,
+                    }
+                ],
+            }
+
+            # Add response_schema if provided
+            if response_format is not None:
+                payload["response_schema"] = response_format.model_json_schema()
+
+            res = requests.post(
+                url=api_url,
+                headers={"Authorization": f"Bearer {host_creds.token}"},
+                json=payload,
+            )
+        except (requests.RequestException, requests.ConnectionError) as e:
+            last_exception = e
+            if attempt < num_retries:
+                _logger.debug(f"Request attempt {attempt + 1} failed with error: {e}")
+                time.sleep(2**attempt)  # Exponential backoff
+                continue
+            else:
+                raise MlflowException(
+                    f"Failed to invoke Databricks model after {num_retries + 1} attempts: {e}",
+                    error_code=INVALID_PARAMETER_VALUE,
+                ) from e
+
+        # Check HTTP status before parsing JSON
+        if res.status_code in [400, 401, 403, 404]:
+            # Don't retry on bad request, unauthorized, not found, or forbidden
+            raise MlflowException(
+                f"Databricks model invocation failed with status {res.status_code}: {res.text}",
+                error_code=INVALID_PARAMETER_VALUE,
+            )
+
+        if res.status_code >= 400:
+            # For other errors, raise exception and potentially retry
+            error_msg = (
+                f"Databricks model invocation failed with status {res.status_code}: {res.text}"
+            )
+            if attempt < num_retries:
+                # Log and retry for transient errors
+                _logger.debug(f"Attempt {attempt + 1} failed: {error_msg}")
+                time.sleep(2**attempt)  # Exponential backoff
+                continue
+            else:
+                raise MlflowException(error_msg, error_code=INVALID_PARAMETER_VALUE)
+
+        # Parse JSON response
+        try:
+            res_json = res.json()
+        except json.JSONDecodeError as e:
+            raise MlflowException(
+                f"Failed to parse JSON response from Databricks model: {e}",
+                error_code=INVALID_PARAMETER_VALUE,
+            ) from e
+
+        # Parse and validate the response using helper function
+        return _parse_databricks_model_response(res_json, res.headers)
+
+    # This should not be reached, but just in case
+    if last_exception:
+        raise MlflowException(
+            f"Failed to invoke Databricks model: {last_exception}",
+            error_code=INVALID_PARAMETER_VALUE,
+        ) from last_exception
+
+
+def _record_judge_model_usage_success_databricks_telemetry(
+    *,
+    request_id: str | None,
+    model_provider: str,
+    endpoint_name: str,
+    num_prompt_tokens: int | None,
+    num_completion_tokens: int | None,
+) -> None:
+    try:
+        from databricks.agents.telemetry import record_judge_model_usage_success
+    except ImportError:
+        _logger.debug(
+            "Failed to import databricks.agents.telemetry.record_judge_model_usage_success; "
+            "databricks-agents needs to be installed."
+        )
+        return
+
+    from mlflow.tracking.fluent import _get_experiment_id
+    from mlflow.utils.databricks_utils import get_job_id, get_job_run_id, get_workspace_id
+
+    record_judge_model_usage_success(
+        request_id=request_id,
+        experiment_id=_get_experiment_id(),
+        job_id=get_job_id(),
+        job_run_id=get_job_run_id(),
+        workspace_id=get_workspace_id(),
+        model_provider=model_provider,
+        endpoint_name=endpoint_name,
+        num_prompt_tokens=num_prompt_tokens,
+        num_completion_tokens=num_completion_tokens,
+    )
+
+
+def _record_judge_model_usage_failure_databricks_telemetry(
+    *,
+    model_provider: str,
+    endpoint_name: str,
+    error_code: str,
+    error_message: str,
+) -> None:
+    try:
+        from databricks.agents.telemetry import record_judge_model_usage_failure
+    except ImportError:
+        _logger.debug(
+            "Failed to import databricks.agents.telemetry.record_judge_model_usage_success; "
+            "databricks-agents needs to be installed."
+        )
+        return
+
+    from mlflow.tracking.fluent import _get_experiment_id
+    from mlflow.utils.databricks_utils import get_job_id, get_job_run_id, get_workspace_id
+
+    record_judge_model_usage_failure(
+        experiment_id=_get_experiment_id(),
+        job_id=get_job_id(),
+        job_run_id=get_job_run_id(),
+        workspace_id=get_workspace_id(),
+        model_provider=model_provider,
+        endpoint_name=endpoint_name,
+        error_code=error_code,
+        error_message=error_message,
+    )
+
+
+@dataclass
+class InvokeJudgeModelHelperOutput:
+    feedback: Feedback
+    model_provider: str
+    model_name: str
+    request_id: str | None
+    num_prompt_tokens: int | None
+    num_completion_tokens: int | None
+
+
+def _invoke_databricks_judge_model(
+    *,
+    model_name: str,
+    prompt: str,
+    assessment_name: str,
+    num_retries: int = 10,
+    response_format: type[pydantic.BaseModel] | None = None,
+) -> InvokeJudgeModelHelperOutput:
+    output = _invoke_databricks_model(
+        model_name=model_name,
+        prompt=prompt,
+        num_retries=num_retries,
+        response_format=response_format,
+    )
+    try:
+        response_dict = json.loads(output.response)
+        feedback = Feedback(
+            name=assessment_name,
+            value=response_dict["result"],
+            rationale=_sanitize_justification(response_dict.get("rationale", "")),
+            source=AssessmentSource(
+                source_type=AssessmentSourceType.LLM_JUDGE,
+                source_id=f"databricks:/{model_name}",
+            ),
+        )
+    except json.JSONDecodeError as e:
+        raise MlflowException(
+            f"Failed to parse the response from the judge. Response: {output.response}",
+            error_code=INVALID_PARAMETER_VALUE,
+        ) from e
+
+    return InvokeJudgeModelHelperOutput(
+        feedback=feedback,
+        model_provider="databricks",
+        model_name=model_name,
+        request_id=output.request_id,
+        num_prompt_tokens=output.num_prompt_tokens,
+        num_completion_tokens=output.num_completion_tokens,
+    )
